Upgrade harvester when energy available
This is a first step towards building in proper 'recycle creeps to
larger versions' code, later ones will swap out the new creep then
send the old one back to a Spawn to be recycled, and limit to only
when the room is full on energy, etc.

A way to more quickly shift from 'bootstrap after creep wipe' back
to 'full production' status, and avoid spawns that are timed close
from depleting overall production because suddenly there's a bunch
of 1-1-1 creeps trying to do everything in an RCL-6 room.